# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: MESH API
  version: 0.3.0
  license: { name: UNLICENSED, identifier: LicenseRef-UNLICENSED }
  description: |
    Contract for the MESH student-collaboration modular monolith. This milestone exposes account,
    profile, self-declared skills, goals, interests, desired skills, weekly availability, GitHub public-repository evidence,
    discovery, connections, project groups, persistent messages, tasks, calendar events, and availability polls.
    GitHub evidence is a limited supporting signal, never proof of expertise. LinkedIn is not collected or scraped.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /projects/{projectId}:
    get:
      operationId: getProject
      summary: Get one project and its members
      parameters: [ { $ref: '#/components/parameters/ProjectId' } ]
      responses:
        '200': { description: Project, content: { application/json: { schema: { $ref: '#/components/schemas/Project' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      operationId: updateProject
      summary: Update a project; owner only
      parameters: [ { $ref: '#/components/parameters/ProjectId' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ProjectInput' } } }
      responses:
        '200': { description: Updated project, content: { application/json: { schema: { $ref: '#/components/schemas/Project' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: deleteProject
      summary: Delete a project; owner only
      parameters: [ { $ref: '#/components/parameters/ProjectId' } ]
      responses:
        '204': { description: Project deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /projects/{projectId}/tasks:
    get:
      operationId: listProjectTasks
      summary: List project tasks; project members only
      parameters: [ { $ref: '#/components/parameters/ProjectId' } ]
      responses:
        '200': { description: Tasks, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/ProjectTask' } } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      operationId: createProjectTask
      summary: Create a task; project members only
      parameters: [ { $ref: '#/components/parameters/ProjectId' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateProjectTaskRequest' } } }
      responses:
        '201': { description: Task created, content: { application/json: { schema: { $ref: '#/components/schemas/ProjectTask' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /projects/{projectId}/tasks/{taskId}:
    put:
      operationId: updateProjectTask
      summary: Update a task; project members only
      parameters: [ { $ref: '#/components/parameters/ProjectId' }, { $ref: '#/components/parameters/TaskId' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/UpdateProjectTaskRequest' } } }
      responses:
        '200': { description: Task updated, content: { application/json: { schema: { $ref: '#/components/schemas/ProjectTask' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: deleteProjectTask
      summary: Delete a task; project members only
      parameters: [ { $ref: '#/components/parameters/ProjectId' }, { $ref: '#/components/parameters/TaskId' } ]
      responses:
        '204': { description: Task deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  responses:
    ValidationError:
      description: Request field validation failed
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unauthorized:
      description: Missing, invalid, or expired bearer token
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    Forbidden:
      description: The signed-in student is not allowed to perform this action
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: The requested resource does not exist or is not accessible
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
  parameters:
...
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    TaskId:
      name: taskId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    ProjectInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 120, example: Campus accessibility map }
        summary: { type: [string, 'null'], maxLength: 1000, example: Map accessible routes around campus. }
        domain: { type: [string, 'null'], maxLength: 80, example: WEB_DEVELOPMENT }
...
    Project:
      type: object
      required: [id, name, ownerUserId, createdAt, members]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        summary: { type: [string, 'null'] }
        domain: { type: [string, 'null'] }
        ownerUserId: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
        members: { type: array, items: { $ref: '#/components/schemas/ProjectMember' } }
    ProjectMember:
      type: object
      required: [userId, username, displayName, role]
      properties:
        userId: { type: string, format: uuid }
        username: { type: string }
        displayName: { type: string }
        avatarKey: { type: [string, 'null'] }
        role: { type: string, enum: [OWNER, MEMBER] }
...
    CreateProjectTaskRequest:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1, maxLength: 160 }
        description: { type: [string, 'null'], maxLength: 2000 }
        assigneeUserId: { type: [string, 'null'], format: uuid }
        dueDate: { type: [string, 'null'], format: date }
    UpdateProjectTaskRequest:
      type: object
      required: [title, status]
      properties:
        title: { type: string, minLength: 1, maxLength: 160 }
        description: { type: [string, 'null'], maxLength: 2000 }
        status: { type: string, enum: [TODO, IN_PROGRESS, DONE] }
        assigneeUserId: { type: [string, 'null'], format: uuid }
        dueDate: { type: [string, 'null'], format: date }
    ProjectTask:
      type: object
      required: [id, projectId, title, status, createdByUserId, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        projectId: { type: string, format: uuid }
        title: { type: string }
        description: { type: [string, 'null'] }
        status: { type: string, enum: [TODO, IN_PROGRESS, DONE] }
        assigneeUserId: { type: [string, 'null'], format: uuid }
        dueDate: { type: [string, 'null'], format: date }
        createdByUserId: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    Error:
      type: object
      required: [timestamp, status, error, message, fields, path]
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer }
        error: { type: string }
        message: { type: string }
        fields: { type: object, additionalProperties: { type: string } }
        path: { type: string }